Pin prompt toolkit to an old version
prompt-toolkit has some implementation bugs where it will print the same
output line multiple times in a row (we can tell because all our log
lines end up duplicated down to the millisecond timestamps), so try to
use a slightly older version with _less_ of the bug, but the bug is
still present.

When the bug starts to happen, you need to restart icli to fix the
internal prompt_toolkit buffer issues.