Fix index option auto-quote adding on order creation
SPX and NDX have "fake underlyings" of SPXW and NDXP on their OCC
symbols for non-monthly variants, but IBKR can't qualify a contract for
SPXW or NDXP directly, so we need to undo the symbol name before
requesting the quote (e.g. when creating a closing side order for an
open position and we're reading the .localSymbol from an exsiting
contract already labeled with SPXW or NDXP).