Add recent EMA log to CLI output rows
This makes the CLI output much more cluttered, but also more useful.

For the equity quotes, we now show:
[longer term EMA] [shorter term EMA] [current price]

For option quotes, we now show:
[shorter term EMA] [current price]

Basically it's useful if you look away from your system for a while,
then come back and want so see a quick price trend (if previous is lower
than current, price is going up; if previous is higher than current,
price is going down, etc).

This means we should do a new console recording for the app/repo
preview.