Improve automated order price handling
Before we were setting limit prices as the DIRECT MIDPOINT of current
bid/ask, but with minor 100 millisecond delays between order calcluation
and order placement and IBKR sending orders to exchanges, we were often
dragging behind the market during rapid price moves.

Now we add more of a price buffer to create a synthetic best-effort (but
still locally price-capped) market orders for when we're trying to
capture momentum.