Only use Order() for orders directly
ib_insync has wrapper classes around Order() for things like
LimitOrder() and MarketOrder(), but Order() is a dataclass while
LimitOrder et al are NOT dataclasses, so the dataclass helpers don't
work on them.

We fix this by ONLY using Order() in our code here everywhere and we
MUST NOT use the wrappers because we expect orders to _always_ be
dataclasses now.