Print rounding result for futures
Some futures have different rounding requirements we can't meet, so we
fake it, but we want to see where the math ended up.

Test case for this failure is /HE rounds to 0.025, but we truncate all
our rounding to 2 places, so we must round /HE to 0.50 instead which
isn't optimal, but the rounding logic is in different places currently.