Enable better prompt-toolkit history recall
`auto_suggest` enables the "history ghost" view when you start to type
where it can be completed with a right arrow.

`enable_history_search` enables the ability to type a partial command
then hit up arrow to cycle through all prefix matches.

the prompt-toolkit documentation says it's not recommended to enable
both of these features at once, but it works for me currently.