Refactor the default symbols provided a little
Not sure why I made SBUX a core symbol to always subscribe on startup.

also stay with SPY and QQQ, if you want IWM and DIA you can add it
yourself.

Also, the Index subscriptions are easier to add on startup then remove
later just to how they are specified.