Add 'range' helper to print OCC strikes to a file
Our algo systems consume OCC strikes and this is an easy way to generate
a wide number of strikes easily.

It's useful having this in icli instead of only the algo agents because
the algo agents only have quote data without knowing the underlying
prices, but here we want to generate a ± point range from the current
prices, so here in icli we have both prices and strike details, so we
can dump it all at once.