Use better event loop task factory on Python 3.12+
This is a more fancy task runner where instead of scheduling an async
call then yielding, the task runner enters the coroutine imediately to
see if it actually does call async itself. If the coroutine doesn't call
async inside for this run, the coroutine returns as normal function and
the task never needed to be scheduled.