Allow option chain lookups to accept quote row IDs
Same as everywhere else... if starts with ':' use the row lookup
behavior instead.

Also we now added an addtional feature where if your requested symbols
ends in "preview" we print all the chains. We need to only show the
chains *sometimes* because this chain-fetch logic runs realtime during
some ranged option order placement too and it's annoying to pollute the
terminal with a month of strikes. Now you can optionally show them or
not per individual request.