Use pre-release pendulum library
For some reasonthe pendulum maintainers refuse to release an updated
real version over the past 2 years, and now all the pip versions can't
install under Python 3.12, so we need to install the beta pendulum
library if we want to use Python 3.12.

Hopefully they get around to updating the official release soon so we
can move this back to regular mainline release versions.