Add client id to order display table
Also refactors the order table creation method a little to be less
indented.

Technically we *can* fetch ALL orders from ALL clients, but when we
fetch ALL orders, it re-submits them when they are fetched (which
doesn't seem great!)

Also we found another "executed orders" API for showing non-open orders
and what happened (placed, canceled) on each of them which would be nice
to report (when we have time to generate a new formatted output
display).