Move per-order log output to individual lines
This was messy showing as a column full of potentially 3, 5, 10 updates
all mashed together. Now each log entry is on a new line per symbol.

I think this can still be cleaned up to look better, but it works for
now and is easier than the previous "20 things in one column" pattern.