Refine `evict` a little better
The evict logic is always tricky because we're basically trying to do a
custom market-with-protection order or even market-to-limit order,
because we never want to flush market depth, but also we don't want
price floating away from us if we're in a rapidly changing phase.

These limits have worked fine for a couple weeks now.

This also fixes the algo name inconsistency due to the other algo
conformity changes a while ago.