Refactor automatic position showing after orders
Logically, this should have worked before, but we're getting stuck in an
async IBKR API update race condition.

So, instead of just running it immediately after the order is sent, wait
a short period of time before attempting to print it?

If this fix is still unreliable, we can move the automatic position
showing to maybe the order update handler itself when it detects a new
order has been executed for a new position the first time?