Improve individual execution trade summary sorts
Same as the execution order summary, but for the individual trade table
too. With trading spreads, all the legs clear at the same timestamp, so
we need to sort more cleverly so the order isn't randomly ordered within
the same second. Now executions nicely group together by spread group
and the legs inside the spread themselves.