Allow `evict` to accept custom algo name
The idea behind `evict` was to have a quick way to just get out of
positions without needing to provide prices, but also without just
running a market order, but then the trick is where do we set the limit
prices? Sometimes in fast rising or falling instruments we get stuck
with a price the market won't meet... so once again we are attempting to
provide better bounds for automatic price setting.

Additionally, now `evict` takes an optional last parameter of a custom
IBKR algo name if you don't want o use the default `evict` chocie algos,
so you _can_ just evict via MKT or MTL or something if you want to
instead of dealing with the price judgements sometimes failing and then
needing to manually hop into `modify` the evict-placed order.

Still TODO: evict shouldn't just fire-and-complete, but it should run a
price tracking algo like `buy` does on purchases (we need to abstract
out the price tracking `buy` logic so it can be used in multiple places
and/or we need to adjust the `buy` command to allow quantity input
instead of only price value input).