Use proper multiplier for execution reporting
Now the 'total' column is the actual cash value and not just
(quantity * price) for instruments having multipliers.

This also fixes the ('total', 'sum') column in the execution-by-order
summary table too.

This commit looks a bit messy though because we also de-indented all the
logic by changing the entire if branch into an early exit check instead.