Workaround IBKR bug not populating spread quotes
Sometimes IBKR just doesn't send data for spread quotes, but if we are
subscribed to quotes for all the legs of the spread, we can just do the
math ourselves and generate a synthetic spread quote anyway.

A spread not having live quote data doesn't impact using the contract to
trade, but trading a contract while having to do the multi-leg addition
and multiplication on your own isn't a good experience.