Improve option chains fetching via automatic dates
This has been annoying since the beginning because IBKR only returns
chains for an entire month, but if it's the last week of a month and you
want the end-of-week chain, you have to get the ENTIRE NEXT MONTH of
chains too.

now instead of manually flipping the code around at end of month to
extend the fetching, we just calculate how many forward months to fetch
when asking for options chains.

This automatically fixes the "end of week is different month from start
of week" and we also now automatically fetch into the next month if we
are past-opex for the current month for more variety in future strikes
we can choose from.