Use "nearest" search instead of bisect for strikes
Fixes an annoyance where we used bisect to find a strike in our sorted
lists of strikes, but it always picks the gap between values without
regard to any difference in values.

This is bad for strikes because, for example, if you have strikes:

100, 105, 110, 115, 120

but your price is 105.31, bisect picks 110 when actually you want the
_closest value_ match not matching by exact position.

Now when we do the bisect, we check the lower value and the upper value
to find if our query is closer to the low or high, then we return the
match with the _smallest_ difference between values (if it's an exact
50% difference to the match it just rounds down for now but maybe
it should be optional or changed to round-up later
(for now, if strikes are 110 and 120 but your query is 105, you get 110).