Remove pygame dependency
Running sounds internal to the async trade API was never 100% reliable
and pygame always had edge cases where it could lock up with 100% CPU
requiring a hard exit to restore the client again (which isn't great)!

Also loading pygame on startup probably added some slowness, so icli may
start a little faster now.

The fix for no pygame sounds is adding externall callouts to awwdio for
trade executions and alerts (which will be re-added in later commits;
currently it is tested as working, but it fires an announcement for
every execution which can be nosiy if you are trading 2,000 shares and
it trades across 20 executions then you get 20 announcements right away,
which probably isn't great? we either need to coalecse them better here
before sending or have awwdio coalece the backlog via a different
tagging mechanism).