Improve/modernize some types and add some timeouts
This also includes the 'awwdio' helper for speaking actions but the code
for using it isn't added yet.

This is mostly just a periodic type cleanup from when we originally
wrote these things before the 3.9+ type improvements.

This also adds some better asyncio timeouts to IBKR API calls because
they can just get stuck for an infinite amount of time if either the
gateway is rebooting or if we are out of API usage credits temporarily
(like requesting too many strikes at once blocks other symbol/contract
lookups... thanks IBKR)