Add automatic straddle command and improve others
Easy instant straddle or strangle with custom width and automatic quantity creation:

Generate a $10,000 debit strangle of strike width 2 from ATM on SPY:

- strad SPY SPY 10_000 2 preview

Generate a $10,000 debit straddle at ATM on SPXW:

- strad SPX SPXW 10_000 0 preview

Technically we should be able to specify delta offsets instead of just
width offsets, but walking the chain width is easier than live delta
calculations currently. Also, technically we should loop this into the
PriceOrQuantity logic so you could specify amounts or contracts, but we
just didn't do it (yet?).

The symbol name is twice because sometimes (like with index options on
NDX/NDXP or SPX/SPXW) the trade symbol (SPXW) isn't the same as the
quote symbol (SPX). If the symbol is the same (SPY) just repeat it both
for the quote and trade symbol.

Other improvements here:

- The `cancel` command now accepts wildcards and multiple symbols
  (`cancel *` drops all your current orders immediately).
- Quote removal command works better with spreads now
- Option chain output is more reliable (we think? need to evaluate if we
  have a stale chain if chains are updated nightly with new strikes...)
- Also just added alias 'ls' to show positions (now 'positions' and 'ls'
  are the same command)