Add generic command expander
This is basically a hack so we can run BUY $X across multiple symbols
without rewriting the buy command.

Example:

> expand buy {AAPL,MSFT} $10_000 AF

Would run 2 async commands from the single 'expand' command:

> buy AAPL $10_000 AF
> buy MSFT $10_000 AF

Also note: the command doesn't divide the quantity or cash amount by the
number of symbols, it just duplicates it for all the combinations of the
other inputs requested.

Also the {A,B,C} or numeric range {3..30} expanders can be at any place
in any command after 'expand' and all combinations of all expansions are
generated.

There's no "preview before running the expanded commands" option, but,
for example with 'buy,' you can use any built-in 'preview' mode of a
command to run them _all_ in preview mode first just to view/verify the
expansions are expanding as expected.