Add a couple more order types (STP/MIT/LIT)
These are currently untested, but it probably works?

We need to find a clean way to expan our order entry process for
accepting the dual price fields for (stop activation price, limit order
price) because right now our 'limit' command only has one price field
which we are using for the stop conditions here (and we are just making
the limit price the stop price which isn't necessarily relaistic... we
will fix eventually).