Fix current price esimation after hours
On weekends and after hours, all the quotes are either -1 or NaN, so
much of our preview/order logic doesn't work.

Now instead of just failing requests with no quotes, we use the last
seen price which should at least be accurate. Also, we use the 'last'
price and not the 'close' price both: prices move after close *and* IBKR
has a problem where, until the next trading day, the 'close' is the
PREVIOUS trading day, so for example over the weekend, every 'close'
price is the THURSDAY close and not FIRDAY close until MONDAY happens...
thanks IBKR.