Fix quote consistency when adding from orders
The `oadd` command automatically adds quotes for all of your open
orders, but we weren't using the shared quote adding abstraction we
centralized a while ago and some of the metadata wasn't being populated
properly.

Now we use the centralized quote adding abstraction so all the extra
fields are populated for things like manual spread math.