Update README with updated relevant details
We should probably have a full documentation system, but this project is
still just me and I only have time for so much {{productivity}} these
days.

At least the commands are mostly self-documenting and the commit history
is the source of truth for detailed reasons about what/why/how/when
things changed over time.