Use proper instance variable for clientId on start
Originally we only had clientId as a global from the environment, but we
wanted the language parser to also have access to the clientId, so now
it's a property of the application and we just need to replace this one
thing here. No functionality changes since this is the same number in
either case.