Avoid sending "last" price instead of live quote
We originally thought the "last" price would be a viable last-traded
price to use in place of bid/ask, but sometimes "last" price is the last
price of the previous day, so it's not useful when trying to price a
LIVE midpoint offset.

We only want to run trade values from live bid/ask midpoints and not
some unrelated "last" value populated before the bid/ask shows up due to
IBKR async APIs filling in each field individually over time after
subscribing.