Add new quote field to options rows for shorting
We have a new field in options rows showing the break-even price if you
short a strike in addition to how much "room" between the current price
and the break even you have on the underlying.

Example:

> SPY   240531P00520000: [u   523.87 (     0.74%)] [iv 0.23] [d -0.12]   ...  (s   519.80 @   4.07)

The (s   519.80 @   4.07)  means "if you short this $520 put strike, your
break even is $519.80 which is $4.07 away from the current underlying"
so your maximum win would be the premium cost if you can survive the
underlying dropping no more than $4.07 before expiration.