Handle lack of audio announcements more gracefully
This code path is only triggered if you launch icli with an environment
variable of `ICLI_AWWDIO_URL`, but it wasn't guarding a case where you
launched with an audio endpoint but the endpoint was down, so you got an
annoying stack trace in your console.

Now when audio is down you just get a clean log line saying audio is
down.