Fix market depth label accuracy
The extra column sort was causing offset problems on the columns
themselves. The goal was to sort the values _inside_ each column, but it
offset the columsn so we ended up with invalid rows like "1 share" but
listed as 5 exchanges, which is impossible because if an exchange is
listed it must contribute at least 1 share to the count.