Improve error message when invalid symbols entered
Previously we printed the entire 100+ line stack trace of the parser
internals when a bad symbol was attempted (like an OCC symbol with not
enough digits), but that was always annoying.

Now we just print the relevant parser error so hopefully it's easier to
see and less polluting of the live console session.