Fix order update for attched order
When using attached orders (like submitting a buy + sell at the same
time), you can't modify the attached order unless you remove the
parent order attachment.

This "fix" causes two changes:

- now you *can* modify attached orders (usually useful after the actual
  order executes and now you want to modify the limit sell price)
- but also now if you attempt to modify an attached order when the
  parent order _has not executed yet_ you'll probably have the order
  canceled because IBKR doesn't allow dual-sided orders? This part is
  untested.

The main benefit here is if you have an attached order sitting idle
_after_ the parent executed, now you can modify it.

Though, if you have a full bracket order of (limit sell, stop sell)
attached to the same parent order, you shouldn't modify either of them
and you just have to cancel them and submit individual orders again (as
far as I can tell... any attempt to modify an order with a parent id
when the parent id has already been executed failed for all my testing).

This update effectively removes the parentid on the order update as far
as I can tell (but because we're submitting a _copy_ of the order to the
update, the original parentId remains in the live order view for now
until the order is actually executed or fully canceled).