Allow closing order with no quantity to imply full
Using the 'limit' menu to generate a closing order, you can use either
-1 quantity to mean "the current entire portfolio position for this
symbol" or you can just leave it blank to get the same effect.

Why leave it blank for full order? It just makes it easier to select the
closing symbol, enter a price, then hit enter a couple times to complete
the menu if you know you want to capture an entire size to exit.