Refactor positions listings code structure
Not sure why we had so much of this inside a "if not self.symbols"
block becuase it's not needed at all?

Also minor logic cleanup by converting the input optional symbols to a
set just to be cleaner.

Also print symbol names being requested for filtering if symbols are
present.

Also fixes the "marketValue" column not having proper commas placed
because it was in the wrong category previously.