Add live ticker quote data to `info` output
The live Ticker object has more metadata than we show in the interface,
so now if you ask for `info SPY` and you are subscribed to SPY quotes,
you now get a fully formatted Ticker object with the live price data
and all your subscribed metadata too.

Also if you ask for `info` on an option symbol, it prints a table of
the 4-way derived greeks data (based on: bid, ask, last, theoretical)
along with some min/max/mean calculations for all the
greeks-per-price-point too.