Audit use contract qualification via cache again
Since the cache used to _not_ work properly, many of these places we
attempted to use the contract cache never showed errors because the
input contracts were always being overwritten by the in/out parameters
themselves.

Now, since the cache is fixed, cached contracts are _returned_ and not
modified in place, so we just had to double check everywhere we call
qualify(*contracts) we assign the return value back to `contracts` again
to use the _actual_ qualified contracts as we expect.