Improve `info` output a little more
Adding spaces between numbers and '%' because it's easier to read with
the space in a terminal font.

Also adding _estimated_ theta decay future prices based on just
multiplying theta by a number of future dates then subtracting the total
decay. This isn't completely accurate because theta decay is non-linear,
but it's good enough for a visual estimation spot check.