Improve greeks reporting without market data
IBKR still provides greeks even if you don't have the underlying market
data available, but previously we conditioned showing greeks only if the
market data was populated.

Now we show greeks if they exist and also we stop defaulting option
ITM/OTM percentage to 0 on no data, but rather just show `nan%` if we
can't do the math due to a missing underlying price in the first place.