Fix automatic spread walking to actually walk
Originally we thought the IBKR "adaptive fast" or "adaptive slow"
algorithms would attempt to "work the bid" and be more effective, but
they actually don't.

"adaptive fast" still grabs the ask if a limit order price is high and
"adaptive slow" mostly passively parks at the bid like a relative order,
but it may wake up to be more aggressive than a resting order sometimes.

So, instead of over-inflating the bid/ask and immediately taking
liquidity from the market, we re-introduce an actual midpoint spread
walker which should work more as expected.