Fix pandas warning in execution report
Apparently it likes to know the format of timestamps before parsing them
or else it gives an error due to running "best guess" attempts on its
own.

These start/end values here are only _times_ like HH:MM:SS so we just
give it a standard time format with no date components for the math
here.