Move "days to expiration" per quote row closer in
This just moves the (d [days]) value next to the (s price @ amt) details
instead of after the duplicate symbol name on the right.

It's nicer to see the days remaining at a glance closer-in on more
narrow terminal windows (or windows with larger fonts which causes the
remainder of the row to be truncated to the right).