Add another quote decimal formatting exception
It's always something around here...

The /LE futures is weird because it has a 40,000 multiplier and is
technically valued at "cents per pound" but everybody quotes it as
"dollars per pound" so the minimum tick is 0.00025 cents per pound, but
when quoted as dollars per pound, the min tick is 0.025 which is 3
decimal places.

The others are for more high-precision (like things trading to 8 decimal
places) instruments which look awful in our current interface, but at
least you can access all the information. I'm not sure our trading
pathway is completely 8-decimal compliant all the way through though.

Also moved a check for 0.01 min tick up to the first case so hopefully
it gets checked first and skips all the other checks most of the time.
Better to have a default case first than as a final fallback since this
thing runs hundreds of times per second.