Improve calculator quote fetching stability
IBKR has so many different ways of saying "no quote data" between 0, -1,
and NaN, so we have to check _all_ of them before using a bid/ask value
anywhere.

This was found because trying to do math against 'SPX' wasn't working
because 'SPX' never has a bid or ask. We always just need to use the
.last value for non-trading index numbers, etc.