Add per-instrument trade tick to detail output
So _this_ is where IBKR hides the details like when options trade $0.01
under $3 then $0.05 over $3... hadn't noticed this data was hiding in
this place before.

We could still optimize these lookups and cache more of these fields so
we don't need network requests for the same data on each detail request,
but the data is at least available and easy to read for now.