Fix mypy misplaced warning
This was in the right place, but a reformat broke the ignore comment
onto a new line which then broke then mypy check again.

If we add the mypy ignore inside the statement line directly instead of
against the close of the statement paren it doesn't get reformatted away
with the current structure.